Demodulator now has it's own spectrum view

+++ src/demod/DemodDefs.h
@@ -71,10 +71,10 @@ class DemodulatorThreadPostIQData: public ReferenceCounter {
 public:
     std::vector<liquid_float_complex> data;
     int bandwidth;
-    float audio_resample_ratio;
+    double audio_resample_ratio;
     msresamp_rrrf audio_resampler;
     msresamp_rrrf stereo_resampler;
-    float resample_ratio;
+    double resample_ratio;
     msresamp_crcf resampler;
 
     DemodulatorThreadPostIQData() :
