+++ src/demod/DemodulatorPreThread.h
@@ -53,11 +53,11 @@ protected:
     AudioThreadInputQueue *audioInputQueue;
 
     msresamp_crcf resampler;
-    float resample_ratio;
+    double resample_ratio;
 
     msresamp_rrrf audio_resampler;
     msresamp_rrrf stereo_resampler;
-    float audio_resample_ratio;
+    double audio_resample_ratio;
 
     DemodulatorThreadParameters params;
     DemodulatorThreadParameters last_params;
